fabric: Define FI_WRITE_COHERENT
MPI defines MPI_WIN_UNIFIED and MPI_WIN_SEPARATE in order
to handle systems that have non-coherent memory between
the host processor and the NIC.  We need to expose that
same capability in libfabric as well.  This allows a NIC
to cache atomic operations or perform RDMA write transfers
to memory that the host processor may not see.

Define an fi_sync mechanism that a process may call to
ensure that all remote writes to its memory space have
been flushed by the NIC and will now be visible to the
process.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>