fabric: Apply FI_CANCEL flag to an endpoint, not per op
Only define FI_CANCEL as a configuration of an endpoint
and not allowed per operation.  If an endpoint has been
configured to support cancelable operations, any operation
may be canceled.

This avoids the need of the provider to check on a per
call basis to see if the call may be canceled.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>