libibverbs: Add libibverbs as a git subtree
Add the libibverbs repository as a subtree.  Provide a
build of libibverbs as part of the libfabric build package.
Update libfabric ibverbs provider to use the libibverbs
header files for its build.  The libibverbs header files
are symbolically linked from the libibverbs subtree into
the include/infiniband directory.

With the upstream version of libibverbs used as a base,
this will make it easier to incorporate any upstream
changes by using the git subtree merge feature.  We
can also support out of tree libibverbs providers.  As
a result, we remove the mlx4 provider from the build.
(Subsequent changes can add it back into the libfabric
package as a subtree).

Signed-off-by: Sean Hefty <sean.hefty@intel.com>