fabric: Streamline atomic calls
Merge all atomic mem and non-mem (write/writemem, etc.) calls.
This reduces the number of calls.  Convert the mem_desc
parameter from a uint64_t to void * to support any sized
local memory key.  This also allows the desc parameter
to be reused for other purposes when local registration
is not needed.

Update the atomic calls to indicate that the provided input
is a count of datatype values, rather than the number of bytes.
Use struct iovec where iov's are referenced.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>