prov/psm: update to the handling of the FI_EVENT flag
The flag has changed from an EP flag to a EP->EQ/EP->counter binding
flag. The flag can now be set separately for differnt EQs and counters
bound to the same EP. This makes it possible to suppress completion
events for send operations while keep the recv operations unaffected.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>