fabric: Support an array of descriptors with sendv and sendmsg
When iovec's are used, each buffer may reference a different
registered memory region.  As a result, we need to support
a different lkey with each region.  Modify the fi_*msg structures
and vectored data transfer operations to take an array of memory
descriptors as input, rather than a single descriptor.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>