prov/sockets: Add sockets provider framework
Create a sockets based provider.  This will allow us to implement
the APIs over any network, so that apps can test that they map
well to their needs.  It can be used for development and
debugging, without requiring the use of specialized hardware.

The provider is not intended to be used for performance.

Provide an initial framework.  Subsequent patches will expand
on the implementation.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>