fabric: Eliminate fi_attr structure
The fi_query and struct fi_attr are somewhat confusing and unnecessary.
The fi_query call requires an fi_info structure as input, which
restricts when it can be used.  The prov_version field is a better fit
with a fabric attribute, and the hw_version and oui are better matches
with domain attributes.  I.e. a provider may support multiple types and
versions of hardware, including supporting hardware from multiple
vendors (e.g. verbs provider).

Rename fi_query to return only the base library version.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>