fabric: Modify EQ readfrom to return fi_addr_t
Update EQ completions to return source addressing data in
the form of fi_addr_t.  This allows the source address to be
used in a response.

To support this change, define that connectionless EPs must
be bound to an AV.  Also allow the app to specify is source
addressing data is required, or if the provider can indicate
if it exists.

The EQ readfrom call is no longer listed as optional, however,
whether source data is available is now endpoint dependent.  When
an event is read from an EQ and no source data is available, the
EQ will return FI_ADDR_UNSPEC.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>