fabric: Define size of tx/rx queues
Define a mechanism that can be used to describe the size of
transmit and receive contexts.  Rather than defining that all
queues contain a fixed number of entries of a fixed size, we
describe a generic set of attributes that allow an app to
determine whether the next operation can be queued with the
tx/rx context.

This supports fixed sized queues, but also allows for queues
that accept a variable number of requests, based on the size
of the request.  There is a restriction that the size of the
underlying queue is a fixed size, in terms of memory.  But
trying to suport a dynamically sizable queue would require
a lot more involvement with the application.

Link from fi_info to the tx/rx attributes, so that the
application can specify the context sizes on endpoint
creation.  This also allows the provider to return the
size information from fi_getinfo.  As part of the change,
move the op_flags from fi_info into the tx/rx attributes,
which will allow the app to specify different flags for
transmit and receive operations.

Update the tostr function accordingly, plus toss in a few
bonus changes to improve code consistency.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>