configury: make FI_<provider>_CONDITIONALS be optional
Still support the per-provider CONDITIONALS m4 macro, but make it
optional: if the provider has it, call it.  But if a provider does not
have it, then just skip it (i.e., it's not an error).

This move is because the only AM_CONDITIONALs we're using right now
are part of the build system and were just slurped up into
fi_provider.m4 (i.e., we don't have any provider-specific
AM_CONDITIONALs in any provider.m4 right now).

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>