fi.h: it is an error if these macros are not already defined
These macros are not for public consumption; they're defined as part
of the build system for libfabric itself.  Hence, if they're not
already defined, it's an error -- they shouldn't be redundantly
defined in Makefile.am *and* fi.h.

As such, make the build system equivalent CPPFLAGS be for the entire
builds (i.e., AM_CPPFLAGS), not just specific to libfabric.la.

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>