fabric: Return open fabric/domain object from fi_getinfo
If an application makes a call to fi_getinfo and an existing
fabric/domain has been opened against the getinfo
output data, return the exising objects to the user.

This simplifies things for the app trying to determine if
it has already opened the returned fabric or domain.  Instead of
needing to search based on the fabric/domain string, it is given
a pointer to the opened objects.  The app is free to ignore the
returned value and open a new object.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>