fabric: Convert CQ len from byte to count
Now that the CQ has been separated from the EQ, it actually makes
more sense to have it accept and return a count of entries found,
rather than the size, since each entry is a fixed size.

This fixes a quirk in the API with fi_cq_readfrom, where len
applies to the buf, but the number of src_addr's is unknown.  Now
the count will apply to both fields.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>