common: Add blocking ring buffer abstraction
Define an abstraction around the ring buffer that supports
a blocking read operation using an fd for blocking/notification.

The same restrictions apply as with the ring buffer.  The
caller must provider serialization and check for overflow.  It
must also ensure that there is data to be read before attempting
to read from the buffer, with the exception of the blocking
sread call.  That call will return the number of bytes read
from the ring.

The implementation supports parallel read and write operations

Signed-off-by: Sean Hefty <sean.hefty@intel.com>