simple: Cleanup examples and increase code sharing
Rename pingpong.c to msg_pingpong.c to for easy identification of EP type.
Combine the rma tests in a single file.

Code cleanups: Move init_test and some prints to a shared file.
Use preprocessor constants for numbers.

Move show_perf to common file.
Adjust performance data for traffic that is not ping-pong in nature.

Signed-off-by: Arun Ilango <arun.ilango@intel.com>
Signed-off-by: Sean Hefty <sean.hefty@intel.com>