prov/psm: use the "service" filed as the port for name resolution
the provider has a built-in simple name server that is enabled
by the SFI_PSM_NAME_SERVER environment variable. It allows a
client to get the remote side PSM epaddr by the node name and
port number. The port number is calculated from the job uuid (by
adding up all the bytes of the uuid and adding 4096 when needed
to make the value greater then 4096). When the service field is
absent, the port number is calculated automatically from the
current job uuid.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>