fabric: Define new shared transmit context object
Shared transmit contexts are currently exposed as struct fi_endpoint.
However, data transfer operations are not posted directely to a
shared tx context.  Instead, transfers are posted to the transport
level endpoints, as those are associated with AVs and CQs.

To avoid confusion, create a new fabric descriptor object for
shared transmit contexts.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>