fabric: separate entry/class for scalable ep
The scalable endpoint is a different type of endpoint.

Scalable endpoints do not support any send/recv
operations. The application is required to create TX/RX
contexts off the scalable endpoint for communication.

Having a separate class and entrypoint for this type of
endpoint will help providers do a better job error-checking,
and simpler usage for apps.

Signed-off-by: Sayantan Sur <sayantan.sur@intel.com>