man: Update documentation to limit connect call to MSG EPs
Connectionless EPs must be bound to an AV and must also be
able to support calling connect. This results in some
confusion regarding what address is passed into connect.

Also, the connected and connectionless data transfer calls
were merged. This reduces the usefulness of calling connect()
on a connectionless EP.

Remove support for connect on connectionless endpoints.  This
simplifies the API.  This is a documentation only change.
The API itself is left unaffected, so that connect could be
added later if needed.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>