fabric: Rename FI_ADDR_UNSPEC -> FI_FORMAT_UNSPEC
It is the address format that is unspecified, not the address.
This change will allow us to use FI_ADDR_UNSPEC for a
different purpose in a subsequent change, where it will mean
that the address was not specified.

As part of this update, the sockets provider is updated to
return ADDR_NOTAVAIL in place of ADDR_UNSPEC where required.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>