fabtest: Add new comprehensive test app framework
Introduce a new test that will be used as a framework
to provide full coverage of the libfabric API.

The test will run as either a persistent service or
single run client app.  The client will read in a test
configuration (currently hard-coded), connect to the
server, and communicate the test parameters to the server.

The client and server will then execute the test, with
the server reporting back the results to the client.
Communication of test data is communicated over standard
TCP sockets.

The framework will be expanded to run multiple tests,
iterating over multiple variables.  It may also be
expanded to run tests using multiple threads and/or
forked processes.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>