fabtests: add "()" to error printfs when identifying a function
Users were getting confused by error messages that looked like this:

   $ simple/fi_msg_pingpong
   fi_getinfo: No such file or directory

The error message was referring to the C function fi_getinfo(), but
there's also an fi_info Linux executable (users' eyes tend to gloss
over the difference between fi_getinfo and fi_info).  This led to
users getting very confused, because they thought that fi_msg_pingpong
was trying to exec fi_info (which would be weird in itself), but then
even though the fi_info exists and is in the user's path, it somehow
wasn't being found.

Clarify the error messages to put "()" when a function is being
discussed.

This was probably only relevant for the fi_getinfo() error messages,
but I went ahead and applied it uniformly throughout the code just so
that it's consistent everywhere.

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>